refactor: 훅분리 , 오타 수정 , 사용법 작성
